+++ NEWS.md
@@ -18,6 +18,7 @@
 - `[`, `[[`, `$`, `$<-`, and `subset()` for `giotto` see `?GiottoClass::subset_giotto`
 - `subset` for `spatIDs()` and `featIDs()`
 - `objName()`, `spatUnit()`, `featType()` generics now return `NA_character_` instead of erroring when used on unsupported classes.
+- `ext()` and `ext<-()` can now be used to get and set extent of `affine2d`
 
 ## new
 - `sliceGiotto()` for pulling out specific spatial units and feature types as independent `giotto` objects